benny you passed

+++ src/main/java/com/qa/math/operations/Adder.java
@@ -2,7 +2,7 @@ package com.qa.math.operations;
 
 public class Adder implements IMathOperation {
 
-	public double mathOperation(int num1, int num2) {
+	public double mathOperation(double num1, double num2) {
 		return num1 + num2;
 	}
