+++ src/main/java/com/qa/math/operations/IMathOperation.java
@@ -1,5 +1,5 @@
 package com.qa.math.operations;
 
 public interface IMathOperation {
-	double mathOperation(int num1, int num2);
+	double mathOperation(double num1, double num2);
 }
